requirements.txt
streamlit==1.31.0
tensorflow==2.18.0
opencv-python-headless==4.11.0.86
Pillow==10.3.0
